+++ src/viola/cl_AC_SGML_txt.h
@@ -5,11 +5,11 @@
 #include "cl_txtDisp.h"
 
 extern ClassInfo class_AC_SGML_txt;
-int helper_AC_SGML_txt_get();
-int helper_AC_SGML_txt_set();
+long int helper_AC_SGML_txt_get();
+long int helper_AC_SGML_txt_set();
 
-int meth_AC_SGML_txt_get();
-int meth_AC_SGML_txt_initialize();
-int meth_AC_SGML_txt_render();
-int meth_AC_SGML_txt_set();
+long int meth_AC_SGML_txt_get();
+long int meth_AC_SGML_txt_initialize();
+long int meth_AC_SGML_txt_render();
+long int meth_AC_SGML_txt_set();
