+++ src/viola/cl_BCard.h
@@ -5,12 +5,12 @@
 #include "cl_field.h"
 
 extern ClassInfo class_BCard;
-int helper_BCard_get();
-int helper_BCard_set();
+long helper_BCard_get();
+long helper_BCard_set();
 
-int meth_BCard_config();
-int meth_BCard_expose();
-int meth_BCard_get();
-int meth_BCard_initialize();
-int meth_BCard_render();
-int meth_BCard_set();
+long meth_BCard_config();
+long meth_BCard_expose();
+long meth_BCard_get();
+long meth_BCard_initialize();
+long meth_BCard_render();
+long meth_BCard_set();
