+++ src/viola/cl_FCard.h
@@ -5,12 +5,12 @@
 #include "cl_pane.h"
 
 extern ClassInfo class_FCard;
-int helper_FCard_get();
-int helper_FCard_set();
+long helper_FCard_get();
+long helper_FCard_set();
 
-int meth_FCard_config();
-int meth_FCard_expose();
-int meth_FCard_get();
-int meth_FCard_initialize();
-int meth_FCard_render();
-int meth_FCard_set();
+long meth_FCard_config();
+long meth_FCard_expose();
+long meth_FCard_get();
+long meth_FCard_initialize();
+long meth_FCard_render();
+long meth_FCard_set();
