+++ src/viola/cl_GIF.h
@@ -5,12 +5,12 @@
 #include "cl_field.h"
 
 extern ClassInfo class_GIF;
-int helper_GIF_get();
-int helper_GIF_set();
+long helper_GIF_get();
+long helper_GIF_set();
 
-int meth_GIF_config();
-int meth_GIF_expose();
-int meth_GIF_get();
-int meth_GIF_initialize();
-int meth_GIF_render();
-int meth_GIF_set();
+long meth_GIF_config();
+long meth_GIF_expose();
+long meth_GIF_get();
+long meth_GIF_initialize();
+long meth_GIF_render();
+long meth_GIF_set();
