+++ src/viola/cl_HTML.h
@@ -5,17 +5,17 @@
 #include "cl_txtDisp.h"
 
 extern ClassInfo class_HTML;
-int helper_HTML_get();
-int helper_HTML_set();
+long helper_HTML_get();
+long helper_HTML_set();
 
-int meth_HTML_back();
-int meth_HTML_config();
-int meth_HTML_get();
-int meth_HTML_initialize();
-int meth_HTML_pathSimplify();
-int meth_HTML_purgeCache();
-int meth_HTML_render();
-int meth_HTML_search();
-int meth_HTML_set();
-int meth_HTML_WWWNameOfFile();
+long meth_HTML_back();
+long meth_HTML_config();
+long meth_HTML_get();
+long meth_HTML_initialize();
+long meth_HTML_pathSimplify();
+long meth_HTML_purgeCache();
+long meth_HTML_render();
+long meth_HTML_search();
+long meth_HTML_set();
+long meth_HTML_WWWNameOfFile();
