+++ src/viola/cl_PS.h
@@ -5,12 +5,12 @@
 #include "cl_field.h"
 
 extern ClassInfo class_PS;
-int helper_PS_get();
-int helper_PS_set();
+long helper_PS_get();
+long helper_PS_set();
 
-int meth_PS_config();
-int meth_PS_get();
-int meth_PS_initialize();
-int meth_PS_info();
-int meth_PS_render();
-int meth_PS_set();
+long meth_PS_config();
+long meth_PS_get();
+long meth_PS_initialize();
+long meth_PS_info();
+long meth_PS_render();
+long meth_PS_set();
