+++ src/viola/cl_TTY.h
@@ -5,9 +5,9 @@
 #include "cl_client.h"
 
 extern ClassInfo class_TTY;
-int helper_TTY_get();
-int helper_TTY_set();
+long helper_TTY_get();
+long helper_TTY_set();
 
-int meth_TTY_get();
-int meth_TTY_set();
-int meth_TTY__startClient();
+long meth_TTY_get();
+long meth_TTY_set();
+long meth_TTY__startClient();
