+++ src/viola/cl_XBM.h
@@ -5,14 +5,14 @@
 #include "cl_field.h"
 
 extern ClassInfo class_XBM;
-int helper_XBM_get();
-int helper_XBM_set();
+long helper_XBM_get();
+long helper_XBM_set();
 
-int meth_XBM_config();
-int meth_XBM_expose();
-int meth_XBM_freeSelf();
-int meth_XBM_get();
-int meth_XBM_initialize();
-int meth_XBM_info();
-int meth_XBM_render();
-int meth_XBM_set();
+long meth_XBM_config();
+long meth_XBM_expose();
+long meth_XBM_freeSelf();
+long meth_XBM_get();
+long meth_XBM_initialize();
+long meth_XBM_info();
+long meth_XBM_render();
+long meth_XBM_set();
