+++ src/viola/cl_XBMBG.h
@@ -5,8 +5,8 @@
 #include "cl_XBM.h"
 
 extern ClassInfo class_XBMBG;
-int helper_XBMBG_get();
-int helper_XBMBG_set();
+long helper_XBMBG_get();
+long helper_XBMBG_set();
 
-int meth_XBMBG_expose();
-int meth_XBMBG_render();
+long meth_XBMBG_expose();
+long meth_XBMBG_render();
