+++ src/viola/cl_XBMButton.h
@@ -5,7 +5,7 @@
 #include "cl_XBM.h"
 
 extern ClassInfo class_XBMButton;
-int helper_XBMButton_get();
-int helper_XBMButton_set();
+long helper_XBMButton_get();
+long helper_XBMButton_set();
 
-int meth_XBMButton_render();
+long meth_XBMButton_render();
