+++ src/viola/cl_XPM.h
@@ -5,13 +5,13 @@
 #include "cl_field.h"
 
 extern ClassInfo class_XPM;
-int helper_XPM_get();
-int helper_XPM_set();
+long helper_XPM_get();
+long helper_XPM_set();
 
-int meth_XPM_config();
-int meth_XPM_freeSelf();
-int meth_XPM_get();
-int meth_XPM_initialize();
-int meth_XPM_info();
-int meth_XPM_render();
-int meth_XPM_set();
+long meth_XPM_config();
+long meth_XPM_freeSelf();
+long meth_XPM_get();
+long meth_XPM_initialize();
+long meth_XPM_info();
+long meth_XPM_render();
+long meth_XPM_set();
