+++ src/viola/cl_client.h
@@ -5,16 +5,16 @@
 #include "cl_field.h"
 
 extern ClassInfo class_client;
-int helper_client_get();
-int helper_client_set();
+long helper_client_get();
+long helper_client_set();
 
-int meth_client_endClient();
-int meth_client_freeSelf();
-int meth_client_get();
-int meth_client_initialize();
-int meth_client_input();
-int meth_client_inputn();
-int meth_client_output();
-int meth_client_set();
-int meth_client_startClient();
+long meth_client_endClient();
+long meth_client_freeSelf();
+long meth_client_get();
+long meth_client_initialize();
+long meth_client_input();
+long meth_client_inputn();
+long meth_client_output();
+long meth_client_set();
+long meth_client_startClient();
