+++ src/viola/cl_glass.h
@@ -5,7 +5,7 @@
 #include "cl_field.h"
 
 extern ClassInfo class_glass;
-int helper_glass_get();
-int helper_glass_set();
+long int helper_glass_get();
+long int helper_glass_set();
