+++ src/viola/cl_hpane.h
@@ -5,13 +5,13 @@
 #include "cl_pane.h"
 
 extern ClassInfo class_hpane;
-int helper_hpane_get();
-int helper_hpane_set();
+long int helper_hpane_get();
+long int helper_hpane_set();
 
-int meth_hpane_config();
-int meth_hpane_get();
-int meth_hpane_initialize();
-int meth_hpane_render();
-int meth_hpane_set();
+long int meth_hpane_config();
+long int meth_hpane_get();
+long int meth_hpane_initialize();
+long int meth_hpane_render();
+long int meth_hpane_set();
