+++ src/viola/cl_pane.h
@@ -5,8 +5,8 @@
 #include "cl_field.h"
 
 extern ClassInfo class_pane;
-int helper_pane_get();
-int helper_pane_set();
+long helper_pane_get();
+long helper_pane_set();
 
 enum PaneConfigTypes {
 	PANE_CONFIG_FREE,
@@ -22,11 +22,11 @@ enum PaneConfigTypes {
 };
 extern char *paneConfigStr[];
 
-int meth_pane_config();
-int meth_pane_clone();
-int meth_pane_clone2();
-int meth_pane_expose();
-int meth_pane_get();
-int meth_pane_initialize();
-int meth_pane_render();
-int meth_pane_set();
+long meth_pane_config();
+long meth_pane_clone();
+long meth_pane_clone2();
+long meth_pane_expose();
+long meth_pane_get();
+long meth_pane_initialize();
+long meth_pane_render();
+long meth_pane_set();
