+++ src/viola/cl_project.h
@@ -5,13 +5,13 @@
 #include "cl_field.h"
 
 extern ClassInfo class_project;
-int helper_project_get();
-int helper_project_set();
+long helper_project_get();
+long helper_project_set();
 
-int meth_project_config();
-int meth_project_expose();
-int meth_project_get();
-int meth_project_initialize();
-int meth_project_info();
-int meth_project_render();
-int meth_project_set();
+long meth_project_config();
+long meth_project_expose();
+long meth_project_get();
+long meth_project_initialize();
+long meth_project_info();
+long meth_project_render();
+long meth_project_set();
