+++ src/viola/cl_radio.h
@@ -5,7 +5,7 @@
 #include "cl_toggle.h"
 
 extern ClassInfo class_radio;
-int helper_radio_get();
-int helper_radio_set();
+long int helper_radio_get();
+long int helper_radio_set();
