+++ src/viola/cl_slider.h
@@ -5,15 +5,15 @@
 #include "cl_field.h"
 
 extern ClassInfo class_slider;
-int helper_slider_get();
-int helper_slider_set();
+long helper_slider_get();
+long helper_slider_set();
 
-int meth_slider_config();
-int meth_slider_processMouseInput();
-int meth_slider_drawLine();
-int meth_slider_drawText();
-int meth_slider_get();
-int meth_slider_initialize();
-int meth_slider_info();
-int meth_slider_render();
-int meth_slider_set();
+long meth_slider_config();
+long meth_slider_processMouseInput();
+long meth_slider_drawLine();
+long meth_slider_drawText();
+long meth_slider_get();
+long meth_slider_initialize();
+long meth_slider_info();
+long meth_slider_render();
+long meth_slider_set();
