+++ src/viola/cl_stack.h
@@ -5,26 +5,26 @@
 #include "cl_pane.h"
 
 extern ClassInfo class_stack;
-int helper_stack_get();
-int helper_stack_set();
+long helper_stack_get();
+long helper_stack_set();
 
-int meth_stack_anyCard();
-int meth_stack_card();
-int meth_stack_config();
-int meth_stack_countBCards();
-int meth_stack_countFCards();
-int meth_stack_create();
-int meth_stack_currentCard();
-int meth_stack_expose();
-int meth_stack_firstCard();
-int meth_stack_get();
-int meth_stack_getCurrentCard();
-int meth_stack_initialize();
-int meth_stack_lastCard();
-int meth_stack_nextCard();
-int meth_stack_previousCard();
-int meth_stack_render();
-int meth_stack_set();
-int meth_stack_setCurrentCard();
+long meth_stack_anyCard();
+long meth_stack_card();
+long meth_stack_config();
+long meth_stack_countBCards();
+long meth_stack_countFCards();
+long meth_stack_create();
+long meth_stack_currentCard();
+long meth_stack_expose();
+long meth_stack_firstCard();
+long meth_stack_get();
+long meth_stack_getCurrentCard();
+long meth_stack_initialize();
+long meth_stack_lastCard();
+long meth_stack_nextCard();
+long meth_stack_previousCard();
+long meth_stack_render();
+long meth_stack_set();
+long meth_stack_setCurrentCard();
 
 VObj *findStackObj();
