+++ src/viola/cl_tray.h
@@ -5,12 +5,12 @@
 #include "cl_field.h"
 
 extern ClassInfo class_tray;
-int helper_tray_get();
-int helper_tray_set();
+long helper_tray_get();
+long helper_tray_set();
 
-int meth_tray_config();
-int meth_tray_get();
-int meth_tray_initialize();
-int meth_tray_render();
-int meth_tray_set();
+long meth_tray_config();
+long meth_tray_get();
+long meth_tray_initialize();
+long meth_tray_render();
+long meth_tray_set();
