+++ src/viola/cl_txt.h
@@ -5,15 +5,15 @@
 #include "cl_pane.h"
 
 extern ClassInfo class_txt;
-int helper_txt_get();
-int helper_txt_set();
+long helper_txt_get();
+long helper_txt_set();
 
-int meth_txt_clone();
-int meth_txt_clone2();
-int meth_txt_config();
-int meth_txt_expose();
-int meth_txt_freeSelf();
-int meth_txt_get();
-int meth_txt_initialize();
-int meth_txt_render();
-int meth_txt_set();
+long meth_txt_clone();
+long meth_txt_clone2();
+long meth_txt_config();
+long meth_txt_expose();
+long meth_txt_freeSelf();
+long meth_txt_get();
+long meth_txt_initialize();
+long meth_txt_render();
+long meth_txt_set();
