+++ src/viola/cl_txtButton.h
@@ -5,11 +5,11 @@
 #include "cl_txt.h"
 
 extern ClassInfo class_txtButton;
-int helper_txtButton_get();
-int helper_txtButton_set();
+long int helper_txtButton_get();
+long int helper_txtButton_set();
 
-int meth_txtButton_get();
-int meth_txtButton_initialize();
-int meth_txtButton_render();
-int meth_txtButton_set();
+long int meth_txtButton_get();
+long int meth_txtButton_initialize();
+long int meth_txtButton_render();
+long int meth_txtButton_set();
