+++ src/viola/cl_txtEdit.h
@@ -5,12 +5,12 @@
 #include "cl_txtDisp.h"
 
 extern ClassInfo class_txtEdit;
-int helper_txtEdit_get();
-int helper_txtEdit_set();
+long int helper_txtEdit_get();
+long int helper_txtEdit_set();
 
-int meth_txtEdit_get();
-int meth_txtEdit_initialize();
-int meth_txtEdit_expose();
-int meth_txtEdit_render();
-int meth_txtEdit_set();
+long int meth_txtEdit_get();
+long int meth_txtEdit_initialize();
+long int meth_txtEdit_expose();
+long int meth_txtEdit_render();
+long int meth_txtEdit_set();
