+++ src/viola/cl_txtEditLine.h
@@ -5,12 +5,12 @@
 #include "cl_txtDisp.h"
 
 extern ClassInfo class_txtEditLine;
-int helper_txtEditLine_get();
-int helper_txtEditLine_set();
+long int helper_txtEditLine_get();
+long int helper_txtEditLine_set();
 
-int meth_txtEditLine_get();
-int meth_txtEditLine_initialize();
-int meth_txtEditLine_expose();
-int meth_txtEditLine_render();
-int meth_txtEditLine_set();
+long int meth_txtEditLine_get();
+long int meth_txtEditLine_initialize();
+long int meth_txtEditLine_expose();
+long int meth_txtEditLine_render();
+long int meth_txtEditLine_set();
