+++ src/viola/cl_txtLabel.h
@@ -5,12 +5,12 @@
 #include "cl_txt.h"
 
 extern ClassInfo class_txtLabel;
-int helper_txtLabel_get();
-int helper_txtLabel_set();
+long int helper_txtLabel_get();
+long int helper_txtLabel_set();
 
-int meth_txtLabel_config();
-int meth_txtLabel_get();
-int meth_txtLabel_initialize();
-int meth_txtLabel_render();
-int meth_txtLabel_set();
+long int meth_txtLabel_config();
+long int meth_txtLabel_get();
+long int meth_txtLabel_initialize();
+long int meth_txtLabel_render();
+long int meth_txtLabel_set();
