+++ src/viola/cl_vpane.h
@@ -5,13 +5,13 @@
 #include "cl_pane.h"
 
 extern ClassInfo class_vpane;
-int helper_vpane_get();
-int helper_vpane_set();
+long int helper_vpane_get();
+long int helper_vpane_set();
 
-int meth_vpane_config();
-int meth_vpane_get();
-int meth_vpane_initialize();
-int meth_vpane_render();
-int meth_vpane_set();
+long int meth_vpane_config();
+long int meth_vpane_get();
+long int meth_vpane_initialize();
+long int meth_vpane_render();
+long int meth_vpane_set();
